Explicitly cast individual un-processable wide characters to integer types before logging them.
In C++20, the underlying std::ostream's operator<< overloads accepting wide character types have been deleted, so the existing code no longer compiles.

This was done to avoid the many accidental cases of logging characters formatted as numbers and strings formatted as pointers, but these cases are actually intentionally logging characters as code points that cannot be interpreted as a character.

PiperOrigin-RevId: 500854154

+++ src/client/client_quality_test_main.cc
@@ -113,8 +113,9 @@ bool GenerateKeySequenceFrom(const std::string& hiragana_sentence,
     } else if (ucs4 == 0x30FB || ucs4 == 0xFF65) {
       key.set_key_code(0x002F);  // "・" -> "/"  "･" -> "/"
     } else {
-      LOG(WARNING) << "Unexpected character: " << std::hex << ucs4 << ": in "
-                   << input << " (" << hiragana_sentence << ")";
+      LOG(WARNING) << "Unexpected character: " << std::hex
+                   << static_cast<uint32_t>(ucs4) << ": in " << input << " ("
+                   << hiragana_sentence << ")";
       return false;
     }
     keys->push_back(key);
